App/requirements.txt
Pillow==7.0.0
requestium==0.1.9
requests==2.22.0
selenium==3.141.0
pdf2image==1.11.0